pre 0036 - fixed dead dogs . added changelog items
0036 - ?
engineer is already 1.2
merged in 1.0.0.4HF2 changes
Skelethon and Ranger are both called Coronus and Necromancer is Called
Skelethon. - fixed
fix coronus dead dogs skill
*check component recipes are they borked?
*cratebox still a problem. fixed - removed the clickable NPC item
(cratebox) And placed the actual npc in the map in the appropriate
location of the outcast hideout entrance. you can use him even if she is
hostile.
*removed unnecessary files from \DAIL\database\DAIL\*
*removed unnecessary *tex files from \*\source\*
*finish the Arlequin and release it
*add/test a pull script to put stuff back into my offline work dir when
i merge something from the wild (if it ever happens)
*add Lumera Relendra credit and npc
*move the occultist and terrorknight component vendors
*fix DGA token icons
*fix hero soul icons
*updated dawn of heroes to 1.4
*check npc tags in DGA
*update vanillabak folder with 1.0.0.4HF2 arz file